requirements.txt
langchain
langchain-community
langchain-openai
langchain-google-genai
langchain-postgres
langchain-text-splitters
pypdf
psycopg[binary,pool]
pgvector
python-dotenv
pydantic-settings
tiktoken
